requirements.txt
Flask
flask-cors
psycopg2
Pillow
transformers
google-generativeai
google-search-results
faiss-cpu
numpy
torch
torchvision
celery
redis
requests
python-dotenv
gevent
